:truck: (frontend): rename signup, signin route
