+++ src/app/plan/[slug]/components/LocationSearch.tsx
@@ -1,0 +1,3 @@
+export default function LocationSearch() {
+  return <div>Location Search</div>
+}
