+++ src/app/signup/components/client/EmailVerification.tsx
@@ -8,7 +8,7 @@ import { useState, useRef } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { toast } from 'react-toastify'
 
-import { type FormValues } from '@/app/sign-up/page'
+import { type FormValues } from '@/app/signup/page'
 
 import ROUTE from '@/constants/route'
 import { verifyCode } from '@/services/api/auth'
